userdefinednetwork: Block net-attach-def deletion when being used
A pod is considered connected to user-defined network when it has
the network-selection annotation specifying the user-defined network.

In order to avoid incomplete state where pods connected to user-
defined-network that doesn't exist, use the finalizer to enable the
controller to ensure the network is not being used before deletion.

Add finalizer to network-attachment-definition of UserDefinedNetwork
instance on creation.

When the UserDefinedNetwork instance is being deleted ensure no pod
is connected to it before removing the finelizer.

Signed-off-by: Or Mergi <ormergi@redhat.com>