udn controller: Simplify reconcile function
Refactor sync CR flow to pass the target namespace name
down the stack to underlying helpers, instead of fetching
the NAD object.

Change the underlying helpers to fetch NAD object reside
in the given namespace.

With this change, the controller reconcile function has less
logic making it easier to follow and maintain.

Signed-off-by: Or Mergi <ormergi@redhat.com>